automator/requirements.txt
yt-dlp>=2026.2.21
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slskd-api==0.2.1
pyyaml>=6.0
spotipy>=2.23.0
python3-discogs-client>=2.7
